requirements.txt
# AI Multi-Agent System - Python Dependencies

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML Framework
langchain==0.0.350
langchain-openai==0.0.2
openai==1.6.1

# Data Processing
pydantic==2.5.0
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Utilities
tiktoken==0.5.2
numpy==1.24.3
pandas==2.0.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: For enhanced features
# streamlit==1.28.1  # Uncomment if you want Streamlit dashboard
# jupyter==1.0.0     # Uncomment for notebook support
# matplotlib==3.7.2  # Uncomment for visualization
# seaborn==0.12.2   # Uncomment for enhanced plots
